Move paging logic to Paging.ky
